new lua supports integers and stores them as long. This is a bit interesting as generally we'd expect a user to have a table with only longs, although maybe they'll put a double in with some wonky math? if a double is in the table there will probably be an InvalidCastOperation, might not be nice for a user if they put in doubles for some reason

+++ src/BizHawk.Client.Common/lua/CommonLibs/JoypadLuaLibrary.cs
@@ -56,7 +56,7 @@ namespace BizHawk.Client.Common
 			var dict = new Dictionary<string, int?>();
 			foreach (var (k, v) in _th.EnumerateEntries<object, object>(controls))
 			{
-				dict[FixString(k.ToString())] = double.TryParse(v.ToString(), out var d) ? (int) d : null;
+				dict[FixString(k.ToString())] = long.TryParse(v.ToString(), out var d) ? (int) d : null;
 			}
 			APIs.Joypad.SetAnalog(dict, controller);
 		}
